requirements.txt
aiohttp>=3.9
pydantic>=2.0
PyYAML>=6.0
lark-oapi>=1.0.0
croniter>=2.0.0
crewai
requests
Pillow
prometheus_client

# 开发与测试依赖
pytest>=8.0
pytest-asyncio>=0.24
pytest-cov>=5.0
aiohttp-test-utils>=0.0.1

